# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "School Management System API",
        "description": "API para la gestión de academias, cursos, matrículas y comunicados",
        "contact": {
            "name": "School Management",
            "email": "info@escuela.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/communications/{id}/send": {
            "post": {
                "tags": [
                    "Communications"
                ],
                "summary": "Enviar un comunicado a los destinatarios",
                "description": "Envía un comunicado existente a todos los tutores asociados con el curso",
                "operationId": "sendCommunication",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del comunicado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Comunicado enviado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Comunicado enviado a 12 destinatarios"
                                        },
                                        "sent": {
                                            "type": "integer",
                                            "example": 12
                                        },
                                        "errors": {
                                            "type": "integer",
                                            "example": 0
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Comunicado no encontrado"
                    },
                    "400": {
                        "description": "El comunicado ya ha sido enviado"
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}